refactor: extract dir-size util, show disk usage in $ and close commands
- Extract getDirSizeBytes/formatBytes to src/utils/dir-size.ts (uses du -sk)
- $ command shows per-dir and total disk usage for sourceWorkingDirs
- close command cleanup summary uses shared util
