+++ src/slack/commands/session-command-handler.ts
@@ -2,6 +2,7 @@ import { CommandHandler, CommandContext, CommandResult, CommandDependencies } fr
 import { CommandParser } from '../command-parser';
 import { userSettingsStore, MODEL_ALIASES, type ModelId } from '../../user-settings-store';
 import { getVerbosityFlags, getVerbosityName, VERBOSITY_NAMES, LOG_DETAIL } from '../output-flags';
+import { getDirSizeBytes, formatBytes as formatBytesUtil } from '../../utils/dir-size';
 
 /**
  * Handles $ prefix commands for current-session-only settings.
@@ -119,12 +120,17 @@ export class SessionCommandHandler implements CommandHandler {
       }
     }
 
-    // Source working dirs
+    // Source working dirs with disk usage
     if (session.sourceWorkingDirs?.length) {
-      lines.push('*Source Working Dirs:*');
+      let totalBytes = 0;
+      const dirLines: string[] = [];
       for (const dir of session.sourceWorkingDirs) {
-        lines.push(`  • \`${dir}\``);
+        const size = getDirSizeBytes(dir);
+        totalBytes += size;
+        dirLines.push(`  • \`${dir}\` — ${formatBytesUtil(size)}`);
       }
+      lines.push(`*Source Working Dirs:* (${formatBytesUtil(totalBytes)} total)`);
+      lines.push(...dirLines);
     }
 
     // Uptime
